requirements.txt
asgiref==3.2.7
beautifulsoup4==4.9.0
bootstrap4==0.1.0
boto3==1.12.41
botocore==1.15.41
Django==3.0.5
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap-form==3.4
django-bootstrap4==1.1.1
django-crispy-forms==1.9.0
django-mathfilters==1.0.0
django-storages==1.9.1
docutils==0.15.2
gunicorn==20.0.4
jmespath==0.9.5
psycopg2==2.8.5
python-dateutil==2.8.1
pytz==2019.3
s3transfer==0.3.3
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
urllib3==1.25.9
whitenoise==5.0.1
